refactor(admin): simplify page structure and adjust CSS padding
Consolidate admin page components by moving `AddAdminDialog` to `page.client.tsx` and removing redundant div structure. Adjust padding values for headings in `globals.css` for better visual consistency.
