+++ app/[locale]/(main)/servers/[id]/admin/page.tsx
@@ -1,12 +1,16 @@
 import { Metadata } from 'next';
 
+import { getCachedServer } from '@/app/[locale]/(main)/servers/[id]/(dashboard)/action';
+import AddAdminDialog from '@/app/[locale]/(main)/servers/[id]/admin/add-admin-dialog';
+
+import Tran from '@/components/common/tran';
+import Divider from '@/components/ui/divider';
+
 import env from '@/constant/env';
 import { isError } from '@/lib/error';
 import { formatTitle, generateAlternate } from '@/lib/utils';
 
-import AddAdminDialog from './add-admin-dialog';
-import PageClient from './page.client';
-import { getCachedServer } from '@/app/[locale]/(main)/servers/[id]/(dashboard)/action';
+import ServerAdminList from './server-admin-list';
 
 type Props = {
 	params: Promise<{ id: string; locale: string }>;
@@ -38,9 +42,15 @@ export default async function Page({ params }: Props) {
 	const { id } = await params;
 
 	return (
-		<div className="flex h-full justify-between flex-col gap-2">
-			<PageClient id={id} />
-			<AddAdminDialog id={id} />
+		<div className="bg-card rounded-md p-4 space-y-2 h-full overflow-hidden flex flex-col">
+			<div className="flex gap-2 justify-between items-center">
+				<h1 className="text-xl">
+					<Tran asChild text="admin" />
+					<AddAdminDialog id={id} />
+				</h1>
+			</div>
+			<Divider />
+			<ServerAdminList id={id} />
 		</div>
 	);
 }
