+++ app/globals.css
@@ -90,15 +90,15 @@
 	}
 
 	h1 {
-		@apply py-4 text-3xl;
+		@apply py-2 text-3xl;
 	}
 
 	h2 {
-		@apply py-2 text-2xl;
+		@apply py-1 text-2xl;
 	}
 
 	h3 {
-		@apply py-1 text-xl;
+		@apply py-0.5 text-xl;
 	}
 
 	h4 {
